# contrato/openapi.json
{
  "components": {
    "schemas": {
...
      "DetalleCliente": {
        "properties": {
          "alertas": {
            "items": {
              "$ref": "#/components/schemas/FilaGestion"
            },
            "title": "Alertas",
            "type": "array"
          },
          "cartera_total": {
            "title": "Cartera Total",
            "type": "string"
          },
          "cliente_nit": {
            "title": "Cliente Nit",
            "type": "string"
          },
          "cliente_nombre": {
            "title": "Cliente Nombre",
            "type": "string"
          },
          "corte": {
            "format": "date",
            "title": "Corte",
            "type": "string"
          },
          "dias_max": {
            "title": "Dias Max",
            "type": "integer"
          },
          "gestiones": {
            "items": {
              "$ref": "#/components/schemas/Gestion"
            },
            "title": "Gestiones",
            "type": "array"
          },
          "marcadores": {
            "items": {
              "type": "string"
            },
            "title": "Marcadores",
            "type": "array"
          },
          "mayor_150": {
            "title": "Mayor 150",
            "type": "string"
          },
          "mayor_90": {
            "title": "Mayor 90",
            "type": "string"
          },
          "n_facturas": {
            "title": "N Facturas",
            "type": "integer"
          },
          "n_vencidas": {
            "title": "N Vencidas",
            "type": "integer"
          },
          "pct_90": {
            "title": "Pct 90",
            "type": "string"
          },
          "pct_vencida": {
            "title": "Pct Vencida",
            "type": "string"
          },
          "por_vencer": {
            "title": "Por Vencer",
            "type": "string"
          },
          "prioridad": {
            "title": "Prioridad",
            "type": "integer"
          },
          "prioridad_etiqueta": {
            "title": "Prioridad Etiqueta",
            "type": "string"
          },
          "vence_hoy": {
            "title": "Vence Hoy",
            "type": "string"
          },
          "vencida": {
            "title": "Vencida",
            "type": "string"
          }
        },
        "required": [
          "cliente_nit",
          "cliente_nombre",
          "corte",
          "cartera_total",
          "por_vencer",
          "vence_hoy",
          "vencida",
          "pct_vencida",
          "mayor_90",
          "pct_90",
          "mayor_150",
          "dias_max",
          "n_facturas",
          "n_vencidas",
          "prioridad",
          "prioridad_etiqueta",
          "marcadores",
          "alertas"
        ],
        "title": "DetalleCliente",
        "type": "object"
      },
...
      "Gestion": {
        "properties": {
          "cliente_nit": {
            "title": "Cliente Nit",
            "type": "string"
          },
          "compromiso_fecha": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compromiso Fecha"
          },
          "compromiso_valor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compromiso Valor"
          },
          "corte": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corte"
          },
          "factura": {
            "title": "Factura",
            "type": "string"
          },
          "fecha": {
            "format": "date-time",
            "title": "Fecha",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "observacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observacion"
          },
          "resultado": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resultado"
          },
          "tipo": {
            "title": "Tipo",
            "type": "string"
          },
          "usuario_id": {
            "title": "Usuario Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "cliente_nit",
          "factura",
          "fecha",
          "corte",
          "usuario_id",
          "tipo",
          "resultado",
          "compromiso_fecha",
          "compromiso_valor",
          "observacion"
        ],
        "title": "Gestion",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ListaGestion": {
        "properties": {
          "corte": {
            "format": "date",
            "title": "Corte",
            "type": "string"
          },
          "filas": {
            "items": {
              "$ref": "#/components/schemas/FilaGestion"
            },
            "title": "Filas",
            "type": "array"
          },
          "pagina": {
            "title": "Pagina",
            "type": "integer"
          },
          "por_pagina": {
            "title": "Por Pagina",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "corte",
          "total",
          "pagina",
          "por_pagina",
          "filas"
        ],
        "title": "ListaGestion",
        "type": "object"
      },
      "NuevaGestion": {
        "description": "Lo que registra el gestor tras contactar al cliente.\n\nEl compromiso de pago es opcional, pero si va, van sus dos partes: media\npromesa no se puede seguir.",
        "properties": {
          "compromiso_fecha": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compromiso Fecha"
          },
          "compromiso_valor": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compromiso Valor"
          },
          "factura": {
            "default": "",
            "description": "Vacio para una alerta de cliente.",
            "title": "Factura",
            "type": "string"
          },
          "observacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observacion"
          },
          "resultado": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resultado"
          },
          "tipo": {
            "description": "llamada, correo, mensaje, visita, acuerdo, disputa u otra",
            "title": "Tipo",
            "type": "string"
          }
        },
        "required": [
          "tipo"
        ],
        "title": "NuevaGestion",
        "type": "object"
      },
      "Panel": {
        "description": "Respuesta del panel: todo lo que §8.1 pide, en una sola llamada.",
        "properties": {
          "aging": {
            "items": {
              "$ref": "#/components/schemas/BarraAging"
            },
            "title": "Aging",
            "type": "array"
          },
          "corte": {
            "format": "date",
            "title": "Corte",
            "type": "string"
          },
          "empresa_id": {
            "title": "Empresa Id",
            "type": "string"
          },
          "generado": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generado"
          },
          "kpis": {
            "items": {
              "$ref": "#/components/schemas/TarjetaKPI"
            },
            "title": "Kpis",
            "type": "array"
          },
          "n_clientes": {
            "title": "N Clientes",
            "type": "integer"
          },
          "n_facturas": {
            "title": "N Facturas",
            "type": "integer"
          },
          "ranking": {
            "items": {
              "$ref": "#/components/schemas/ClienteEnRanking"
            },
            "title": "Ranking",
            "type": "array"
          },
          "reglas_inactivas": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Reglas Inactivas",
            "type": "object"
          },
          "version_parametros": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Parametros"
          }
        },
        "required": [
          "empresa_id",
          "corte",
          "kpis",
          "aging",
          "ranking",
          "n_clientes",
          "n_facturas"
        ],
        "title": "Panel",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "\nMotor de alertas de cartera de BUSINT.\n\nConvierte el aging de cuentas por cobrar en una lista de trabajo priorizada.\nToda alerta lleva la explicacion de por que se disparo: la regla, el parametro\nvigente y el valor que la disparo.\n\n**Los montos viajan como cadena**, no como numero, para conservar los dos\ndecimales sin el error de redondeo de los flotantes de JavaScript.\n\nAutenticacion: `POST /api/v1/sesion` devuelve un token; el resto de endpoints\nlo esperan en `Authorization: Bearer`. Los permisos son acumulativos segun el\nrol (Consulta, Gestor, Coordinador, Administrador).\n",
    "title": "BUSINT - Motor de alertas de cartera",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/clientes/{cliente_nit}": {
      "get": {
        "operationId": "detalle_api_v1_clientes__cliente_nit__get",
        "parameters": [
          {
            "in": "path",
            "name": "cliente_nit",
            "required": true,
            "schema": {
              "title": "Cliente Nit",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "corte",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Corte"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleCliente"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detalle del cliente",
        "tags": [
          "gestion"
        ]
      }
    },
    "/api/v1/clientes/{cliente_nit}/gestiones": {
      "get": {
        "operationId": "gestiones_api_v1_clientes__cliente_nit__gestiones_get",
        "parameters": [
          {
            "in": "path",
            "name": "cliente_nit",
            "required": true,
            "schema": {
              "title": "Cliente Nit",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Gestion"
                  },
                  "title": "Response Gestiones Api V1 Clientes  Cliente Nit  Gestiones Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Historial de gestiones del cliente",
        "tags": [
          "gestion"
        ]
      },
      "post": {
        "description": "§11: el gestor registra la gestion y la alerta pasa a gestionada.\n\nNo toca el saldo ni el bucket. §16 separa el estado de la gestion del de la\nfactura: una factura sigue vencida aunque ya se haya gestionado.",
        "operationId": "registrar_gestion_api_v1_clientes__cliente_nit__gestiones_post",
        "parameters": [
          {
            "in": "path",
            "name": "cliente_nit",
            "required": true,
            "schema": {
              "title": "Cliente Nit",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "corte",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Corte"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NuevaGestion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gestion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Registrar una gestion de cobranza",
        "tags": [
          "gestion"
        ]
      }
    },
...
    "/api/v1/gestion": {
      "get": {
        "operationId": "lista_api_v1_gestion_get",
        "parameters": [
          {
            "in": "query",
            "name": "prioridad_minima",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 4,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prioridad Minima"
            }
          },
          {
            "in": "query",
            "name": "bucket",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bucket"
            }
          },
          {
            "in": "query",
            "name": "vendedor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vendedor"
            }
          },
          {
            "in": "query",
            "name": "zona",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zona"
            }
          },
          {
            "in": "query",
            "name": "estado",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "activa",
              "title": "Estado"
            }
          },
          {
            "description": "NIT o numero de factura",
            "in": "query",
            "name": "busqueda",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "NIT o numero de factura",
              "title": "Busqueda"
            }
          },
          {
            "in": "query",
            "name": "orden",
            "required": false,
            "schema": {
              "default": "prioridad",
              "pattern": "^(prioridad|dias|saldo|cliente)$",
              "title": "Orden",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "pagina",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Pagina",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "por_pagina",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Por Pagina",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "corte",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Corte"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListaGestion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lista de gestion",
        "tags": [
          "gestion"
        ]
      }
    },
    "/api/v1/panel": {
      "get": {
        "operationId": "panel_api_v1_panel_get",
        "parameters": [
          {
            "in": "query",
            "name": "corte",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Corte"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Panel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Panel de control",
        "tags": [
          "panel"
        ]
      }
    },
    "/api/v1/panel/criticos": {
      "get": {
        "operationId": "criticos_api_v1_panel_criticos_get",
        "parameters": [
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limite",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "corte",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Corte"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClienteEnRanking"
                  },
                  "title": "Response Criticos Api V1 Panel Criticos Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clientes que requieren atencion",
        "tags": [
          "panel"
        ]
      }
    },
...
  },
...
}